[Wallet] Move 7702 batch transaction test to integration tests
